refactor: update import paths for MovieCarousel, LoadingSpinner, SearchBar, and SearchSuggestions components to core directory, improving code organization and maintainability
